+++ Classes/Service/DatabaseService.php
@@ -13,14 +13,18 @@ use T3v\T3vCore\Service\AbstractService;
 class DatabaseService extends AbstractService {
   /**
    * Setup the Eloquent ORM.
+   *
+   * @param string $connection The optional connection, defaults to `Default`
    */
-  public static function setup() {
+  public static function setup($connection = 'Default') {
+    $connection = (string) $connection;
+
     // First, create a new `Capsule` manager instance. Capsule aims to make configuring the library for usage outside of
     // the Laravel framework as easy as possible.
     $capsule = new Capsule();
 
     // Add connection to Capsule.
-    $capsule->addConnection(self::getConnection());
+    $capsule->addConnection(self::getConnection($connection));
 
     // Make this Capsule instance available globally via static methods.
     $capsule->setAsGlobal();
@@ -32,39 +36,29 @@ class DatabaseService extends AbstractService {
   /**
    * Returns the connection configuration from the Global TYPO3 configuration options.
    *
+   * @param string $connection The optional connection, defaults to `Default`
    * @return array The connection configuration
    * @link https://laravel.com/docs/master/database#configuration The Laravel Database Configuration
    * @link https://docs.typo3.org/typo3cms/CoreApiReference/stable/ApiOverview/GlobalValues/GlobalVariables The TYPO3 Global variables
    */
-  protected static function getConnection() {
-    $extensionConfiguration = unserialize($GLOBALS['TYPO3_CONF_VARS']['EXT']['extConf']['t3v_datamapper']);
+  protected static function getConnection($connection = 'Default') {
+    $configuration = [];
 
-    $connection = [];
+    if (is_array($GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'][$connection])) {
+      $extensionConfiguration = unserialize($GLOBALS['TYPO3_CONF_VARS']['EXT']['extConf']['t3v_datamapper']);
 
-    if (is_array($GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'])) {
-      $connection = [
+      $configuration = [
         'driver'    => $extensionConfiguration['driver'] ?: 'mysql',
-        'host'      => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections']['Default']['host'],
-        'username'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections']['Default']['user'],
-        'password'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections']['Default']['password'],
-        'database'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections']['Default']['dbname'],
-        'charset'   => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections']['Default']['charset'] ?: 'utf8',
-        'collation' => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections']['Default']['collation'] ?: 'utf8_general_ci',
-        'prefix'    => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections']['Default']['prefix'] ?: ''
-      ];
-    } else {
-      $connection = [
-        'driver'    => $extensionConfiguration['driver'] ?: 'mysql',
-        'host'      => $GLOBALS['TYPO3_CONF_VARS']['DB']['host'],
-        'username'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['username'],
-        'password'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['password'],
-        'database'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['database'],
-        'charset'   => $GLOBALS['TYPO3_CONF_VARS']['DB']['charset'] ?: 'utf8',
-        'collation' => $GLOBALS['TYPO3_CONF_VARS']['DB']['collation'] ?: 'utf8_general_ci',
-        'prefix'    => $GLOBALS['TYPO3_CONF_VARS']['DB']['prefix'] ?: ''
+        'host'      => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'][$connection]['host'],
+        'username'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'][$connection]['user'],
+        'password'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'][$connection]['password'],
+        'database'  => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'][$connection]['dbname'],
+        'charset'   => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'][$connection]['charset'] ?: 'utf8',
+        'collation' => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'][$connection]['collation'] ?: 'utf8_general_ci',
+        'prefix'    => $GLOBALS['TYPO3_CONF_VARS']['DB']['Connections'][$connection]['prefix'] ?: ''
       ];
     }
 
-    return $connection;
+    return $configuration;
   }
 }