+++ Scripts/Setup.sh
@@ -16,3 +16,6 @@ composer require typo3/cms="$TYPO3_VERSION"
 
 # Reset the changes
 git checkout composer.json
+
+# Try to keep environment pollution down, EPA loves us
+unset TYPO3_VERSION