+++ ext_localconf.php
@@ -3,11 +3,9 @@ defined('TYPO3_MODE') or die();
 
 // === Variables ===
 
-$namespace          = 't3v';
-$extensionKey       = $_EXTKEY;
-$extensionSignature = \TYPO3\CMS\Core\Utility\GeneralUtility::underscoredToUpperCamelCase($namespace . '.' . $extensionKey);
+$extensionKey = $_EXTKEY;
 
-// === Extbase Commands ===
+// === Extbase Command Controllers ===
 
 if (TYPO3_MODE === 'BE') {
   $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['extbase']['commandControllers'][$extensionKey] = \T3v\T3vDataMapper\Command\PageLanguageOverlayCommandController::class;